fix(hubspot): fold token-refresh retry tests into test_hubspot_auth
Master's #63732 added test_hubspot_auth.py for the same refresh function. This branch's separate test_auth.py both duplicated those tests and collided at collection time with sources/salesforce/test/test_auth.py (same basename, no __init__.py packages), failing the Temporal test run with a pytest collection error.

Remove test_auth.py and move its distinct retry-behaviour cases (retry-then-succeed with call-count assertions, exhaustion, no-retry on non-transient) into test_hubspot_auth.py. Also add an autouse fixture there that neutralises tenacity's backoff sleep and restores it afterwards — now that hubspot_refresh_access_token is @retry-decorated, the existing transient-status tests would otherwise perform real exponential-backoff sleeps.

Generated-By: PostHog Code
Task-Id: 6977547a-5fb1-4de3-af13-c39820bb2d74
